requirements.txt
django==4.2.16
social-auth-app-django==5.4.1
python-jose[cryptography]==3.3.0
django-environ==0.11.2
gunicorn==22.0.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
django-lockdown==4.0.0
pillow==10.3.0
crispy-bootstrap5==2024.2
django-crispy-forms==2.2
django-select2==8.2.0
django-recaptcha==4.0.0
pandas==2.2.2
django-filter==24.3
django-easy-audit==1.3.6